Create zxx.afsh4n-ddos
Disclaimer: I am not responsible for any misuse of this code. Please use this repository responsibly and within the bounds of the law, This repository is for educational purposes only. I do not condone or encourage any illegal activities, including DDOS attacks, This repository is licensed under the MIT License.